+++ packages/oidc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vercel/oidc
 
+## 1.0.1
+
+### Patch Changes
+
+- Added refresh token support ([#13608](https://github.com/vercel/vercel/pull/13608))
+- - Optionally pass "true" to getVercelOidcToken() to prompt the function to refresh the token if it is expired or missing
+- - Depends on vercel cli being installed and properly linked with the project
+
 ## 1.0.0
 
 ### Major Changes